requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.7
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-htmlmin==0.11.0
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
whitenoise==6.6.0
